Moved to a Debugger class
